# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ObjectLiteral": {
				"description": "Interface of the simple literal object with any string keys.",
				"properties": {},
				"type": "object",
				"additionalProperties": {}
			},
			"Scores": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDated": {
						"type": "string",
						"format": "date-time"
					},
					"total_score": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"createdDate",
					"updatedDated",
					"total_score",
					"category",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDated": {
						"type": "string",
						"format": "date-time"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"scores": {
						"items": {
							"$ref": "#/components/schemas/Scores"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdDate",
					"updatedDated",
					"firstName",
					"lastName",
					"email",
					"password",
					"scores"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "YamdGame",
		"version": "0.0.1",
		"description": "Awesome project developed with TypeORM.",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ObjectLiteral"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"properties": {
												"updatedDated": {
													"type": "string",
													"format": "date-time"
												},
												"createdDate": {
													"type": "string",
													"format": "date-time"
												},
												"id": {
													"type": "number",
													"format": "double"
												},
												"scores": {
													"items": {
														"$ref": "#/components/schemas/Scores"
													},
													"type": "array"
												},
												"password": {
													"type": "string"
												},
												"email": {
													"type": "string"
												},
												"lastName": {
													"type": "string"
												},
												"firstName": {
													"type": "string"
												}
											},
											"required": [
												"updatedDated",
												"createdDate",
												"id",
												"scores",
												"password",
												"email",
												"lastName",
												"firstName"
											],
											"type": "object"
										},
										{
											"$ref": "#/components/schemas/ObjectLiteral"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ObjectLiteral"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}